modules/revhumanize/requirements.txt
# FastAPI & Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# API & Networking
httpx==0.26.0
requests==2.31.0

# Data & Validation
pydantic==2.5.3
python-dotenv==1.0.0

# Async
asyncio==3.4.3
aiofiles==23.2.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
celery[redis]==5.3.4
flower==2.0.1
redis==4.5.4
python-jose[cryptography]==3.3.0
